+++ core-mvp-binding/RELEASE_NOTES.md
@@ -5,6 +5,8 @@
 * ANDDEP-687 Added "androidx.constraintlayout:constraintlayout" dependency with "implementation" type
 * Clearing `BaseRxFragmentView.viewDisposable` move from onDestroy to onDestroyView
 * ANDDEP-671 Renamed loadable to response
+* AndroidSchedulers.mainThread() replaced with ImmediateScheduler 
+ in subscription methods of BaseRxViews.
 ## 0.3.0
 ##### Core-mvp-binding
 * Renamed `onViewDetached ()` -> `onViewDetach ()`